FEATURE : 포인트 인출 기능 추가 #228
마이페이지 포인트 인출 API

closed #228

+++ src/components/units/users/sales/sales.queries.ts
@@ -4,7 +4,7 @@ export const FETCH_ORDER_BY_CREATOR = gql`
   query fetchOrdersByCreatorId(
     $startDate: DateTime
     $endDate: DateTime
-    $page: Float!
+    $page: Int!
   ) {
     fetchOrdersByCreatorId(
       startDate: $startDate
